NEW: Cards data displayed !!!! woohoooo !!! Calender and date picking data Snackbar introduced
Bugs:
Checkbox returning false/true always

Solved:
Month updating bug, Year skipping to september bug solved.
New groupImage picker sending data bug solved.
Group data populating list view, Group pushing data.
Onclick listener added to RecyclerView Adapter class, Login done, logout done.
 Profile edit done, Create group picture upload done.
Basic chatroom done, logout error solved.
Trial adapter class getting data.
Creating groups bug solved
