+++ app/src/main/AndroidManifest.xml
@@ -46,11 +46,6 @@
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".activities.GroupDetailsActivity" />
         <activity
-            android:name=".activities.NextBuildActivity"
-            android:parentActivityName=".activities.LoginActivity"
-            android:screenOrientation="portrait"
-            android:theme="@style/AppTheme.NoActionBar" />
-        <activity
             android:name=".Main2Activity"
             android:label="@string/title_activity_main2"
             android:theme="@style/AppTheme.NoActionBar" />
