+++ app/src/main/res/layout/group_view.xml
@@ -32,7 +32,8 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:id="@+id/group_key"
-            android:text="sample"/>
+            android:text="sample"
+            android:visibility="gone"/>
 
         <TextView
         android:id="@+id/text_last_message"
